requirements.txt
ben_cogs[sql]>=0.13.0,<1.0.0
discord.py>=1.2.2,<2.0.0
jishaku>=1.14.0,<2.0.0
json5
parsedatetime
python-dateutil
